+++ src/main/java/dk/aau/cs/ds306e18/tournament/model/PsyonixBotFromConfig.java
@@ -3,11 +3,13 @@ package dk.aau.cs.ds306e18.tournament.model;
 import dk.aau.cs.ds306e18.tournament.rlbot.configuration.BotSkill;
 import dk.aau.cs.ds306e18.tournament.rlbot.configuration.BotType;
 
+import java.io.IOException;
+
 public class PsyonixBotFromConfig extends BotFromConfig {
 
     private BotSkill skill;
 
-    public PsyonixBotFromConfig(String pathToConfig, BotSkill skill) {
+    public PsyonixBotFromConfig(String pathToConfig, BotSkill skill) throws IOException {
         super(pathToConfig);
         this.skill = skill;
     }