+++ src/main/java/dk/aau/cs/ds306e18/tournament/rlbot/configuration/BotSkill.java
@@ -1,29 +1,37 @@
 package dk.aau.cs.ds306e18.tournament.rlbot.configuration;
 
-/**
- * Enum for bot skill values in the rlbot.cfg
- */
+import rlbot.flat.PsyonixSkill;
+
 public enum BotSkill {
-    ROOKIE("0.0"),
-    PRO("0.5"),
-    ALLSTAR("1.0");
+    ROOKIE("Rookie", PsyonixSkill.Rookie),
+    PRO("Pro", PsyonixSkill.Pro),
+    ALLSTAR("AllStar", PsyonixSkill.AllStar);
 
-    private final String configValue;
+    private final String name;
+    private final int value;
 
-    BotSkill(String configValue) {
-        this.configValue = configValue;
+    BotSkill(String name, int value) {
+        this.name = name;
+        this.value = value;
     }
 
-    public String getConfigValue() {
-        return configValue;
+    @Override
+    public String toString() {
+        return name;
     }
 
-    /**
-     * Turn a number into a valid skill type
-     */
-    public static BotSkill getSkillFromNumber(double skill) {
-        if (skill <= 0.0) return ROOKIE;
-        if (skill <= 0.5) return PRO;
+    public String getName() {
+        return name;
+    }
+
+    public int getValue() {
+        return value;
+    }
+
+    public static BotSkill get(String name) {
+        for (BotSkill m : values()) {
+            if (m.name.equals(name)) return m;
+        }
         return ALLSTAR;
     }
 }