+++ src/main/java/dk/aau/cs/ds306e18/tournament/ui/RLBotSettingsTabController.java
@@ -1,7 +1,7 @@
 package dk.aau.cs.ds306e18.tournament.ui;
 
 import dk.aau.cs.ds306e18.tournament.model.Tournament;
-import dk.aau.cs.ds306e18.tournament.rlbot.MatchRunner;
+import dk.aau.cs.ds306e18.tournament.rlbot.MatchControl;
 import dk.aau.cs.ds306e18.tournament.rlbot.RLBotSettings;
 import dk.aau.cs.ds306e18.tournament.rlbot.configuration.MatchConfig;
 import dk.aau.cs.ds306e18.tournament.rlbot.configuration.MatchConfigOptions.*;
@@ -31,10 +31,9 @@ public class RLBotSettingsTabController {
     public CheckBox instantStartCheckbox;
     public CheckBox writeOverlayDataCheckbox;
     public TextField overlayPathTextField;
+    public CheckBox useGreenScreen;
     public Button chooseOverlayPathButton;
-    public CheckBox useRLBotGUIPythonCheckbox;
     public Button rlbotRunnerOpenButton;
-    public Button rlbotRunnerCloseButton;
     public Button rlbotRunnerStopMatchButton;
     public ChoiceBox<MatchLength> matchLengthChoiceBox;
     public ChoiceBox<MaxScore> maxScoreChoiceBox;
@@ -67,26 +66,11 @@ public class RLBotSettingsTabController {
         // General match settings
         setupChoiceBox(gameMapChoiceBox, GameMap.values(), matchConfig.getGameMap(), MatchConfig::setGameMap);
         setupChoiceBox(gameModeChoiceBox, GameMode.values(), matchConfig.getGameMode(), MatchConfig::setGameMode);
-        skipReplaysCheckbox.setSelected(matchConfig.isSkipReplays());
-        skipReplaysCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().getMatchConfig().setSkipReplays(newValue);
-        });
-        instantStartCheckbox.setSelected(matchConfig.isInstantStart());
-        instantStartCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().getMatchConfig().setInstantStart(newValue);
-        });
-        renderingCheckbox.setSelected(matchConfig.isRenderingEnabled());
-        renderingCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().getMatchConfig().setRenderingEnabled(newValue);
-        });
-        stateSettingCheckbox.setSelected(matchConfig.isStateSettingEnabled());
-        stateSettingCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().getMatchConfig().setStateSettingEnabled(newValue);
-        });
-        autoSaveReplaysCheckbox.setSelected(matchConfig.isAutoSaveReplays());
-        autoSaveReplaysCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().getMatchConfig().setAutoSaveReplays(newValue);
-        });
+        setupCheckBox(skipReplaysCheckbox, matchConfig.doSkipReplays(), MatchConfig::setSkipReplays);
+        setupCheckBox(instantStartCheckbox, matchConfig.isInstantStart(), MatchConfig::setInstantStart);
+        setupCheckBox(renderingCheckbox, matchConfig.isRenderingEnabled(), MatchConfig::setRenderingEnabled);
+        setupCheckBox(stateSettingCheckbox, matchConfig.isStateSettingEnabled(), MatchConfig::setStateSettingEnabled);
+        setupCheckBox(autoSaveReplaysCheckbox, matchConfig.doAutoSaveReplays(), MatchConfig::setAutoSaveReplays);
 
         // Mutators
         setupChoiceBox(matchLengthChoiceBox, MatchLength.values(), matchConfig.getMatchLength(), MatchConfig::setMatchLength);
@@ -109,7 +93,7 @@ public class RLBotSettingsTabController {
         boolean writeOverlay = settings.writeOverlayDataEnabled();
         writeOverlayDataCheckbox.setSelected(writeOverlay);
         writeOverlayDataCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().setWriteOverlayData(newValue);
+            settings.setWriteOverlayData(newValue);
             overlayPathTextField.setDisable(!newValue);
             chooseOverlayPathButton.setDisable(!newValue);
         });
@@ -122,23 +106,31 @@ public class RLBotSettingsTabController {
                 settings.setOverlayPath(path);
             }
         });
-
-        useRLBotGUIPythonCheckbox.setSelected(settings.useRLBotGUIPythonIfAvailable());
-        useRLBotGUIPythonCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
-            Tournament.get().getRlBotSettings().setUseRLBotGUIPythonIfAvailable(newValue);
+        setupCheckBox(useGreenScreen, Tournament.get().getUseGreenScreen(), (conf, val) -> {
+            Tournament.get().setUseGreenScreen(val);
         });
 
         update();
     }
 
     /**
-     * Setup a choice box for match config options.
+     * Set up a choice box for match config options.
      */
     private <T> void setupChoiceBox(ChoiceBox<T> choiceBox, T[] values, T current, BiConsumer<MatchConfig, T> onSelect) {
         choiceBox.setItems(FXCollections.observableArrayList(values));
         choiceBox.getSelectionModel().select(current);
         choiceBox.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
             onSelect.accept(Tournament.get().getRlBotSettings().getMatchConfig(), newValue);
+        });
+    }
+
+    /**
+     * Set up a checkbox for match config options.
+     */
+    private void setupCheckBox(CheckBox checkBox, boolean current, BiConsumer<MatchConfig, Boolean> onSelect) {
+        checkBox.setSelected(current);
+        checkBox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+            onSelect.accept(Tournament.get().getRlBotSettings().getMatchConfig(), checkBox.isSelected());
         });
     }
 
@@ -156,7 +148,7 @@ public class RLBotSettingsTabController {
         // General match settings
         gameMapChoiceBox.getSelectionModel().select(matchConfig.getGameMap());
         gameModeChoiceBox.getSelectionModel().select(matchConfig.getGameMode());
-        skipReplaysCheckbox.setSelected(matchConfig.isSkipReplays());
+        skipReplaysCheckbox.setSelected(matchConfig.doSkipReplays());
         instantStartCheckbox.setSelected(matchConfig.isInstantStart());
 
         // Mutators
@@ -181,15 +173,11 @@ public class RLBotSettingsTabController {
     }
 
     public void onActionRLBotRunnerOpen(ActionEvent actionEvent) {
-        MatchRunner.startRLBotRunner();
-    }
-
-    public void onActionRLBotRunnerClose(ActionEvent actionEvent) {
-        MatchRunner.closeRLBotRunner();
+        MatchControl.get().launchConnectAndRunRLBotIfNeeded();
     }
 
     public void onActionRLBotRunnerStopMatch(ActionEvent actionEvent) {
-        MatchRunner.stopMatch();
+        MatchControl.get().stopMatch();
     }
 
     public void onActionChooseOverlayPath(ActionEvent actionEvent) {